// web/web-app/src/app/ingredients/types/IngredientFacts.ts
export interface IngredientFacts {
  calories: number | undefined;
  carbs: number | undefined;
  sugars: number | undefined;
  protein: number | undefined;
  totalFat: number | undefined;
  saturatedFat: number | undefined;
  monounsaturatedFat: number | undefined;
  polyunsaturatedFat: number | undefined;
}